Genesis Staff Listing Update
Updated GSL to work for Genesis 2.0 themes using HTML 5. Added an
option to allow the user to specify the page that should use the
page-gslstaff.php page template. Modified the plugin so that the
page-gslstaff.php page template actually gets loaded from the plugin.
Tried to remove all references to the technology theme that this plugin
is based on.